Make it easier to make a silenced test deployment
This prefixes all the notification commands by $USER6$. By defining $USER6$
in your resource.cfg to be a command that writes to a log file instead of
running the command passed on the command-line, you can debug the config
without sending a bunch of spurious messages to users. OTOH, if you leave
it empty, nagios will work as expected.